requirements.txt
amqp==2.6.0
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
billiard==3.6.3.0
blosc==1.9.1
bottle==0.12.17
celery==4.4.5
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
colorlog==4.0.2
compress==0.0.2
coverage==4.5.4
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.4
django-appconf==1.0.3
django-autoslug==1.9.6
django-celery-results==1.2.1
django-ckeditor==5.9.0
django-compressor==2.3
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-formtools==2.1
django-haystack==2.8.1
django-import-export==2.3.0
django-ipware==2.1.0
django-js-asset==1.2.2
django-mdeditor==0.1.16
django-ranged-response==0.2.0
django-redis==4.11.0
django-redis-cache==2.1.0
django-reversion==3.0.7
django-simple-captcha==0.5.12
django-tinymce==3.0.2
django-uuslug==1.1.8
djangorestframework==3.11.0
djangorestframework-stubs==1.2.0
dnspython==1.16.0
docopt==0.6.2
elasticsearch==7.0.0
elasticsearch-dsl==7.0.0
et-xmlfile==1.0.1
eventlet==0.25.2
filetype==1.0.5
future==0.15.2
greenlet==0.4.16
har2case==0.3.1
httplib2==0.9.2
httprunner==2.3.3
idna==2.8
importlib-metadata==1.6.1
ipaddress==1.0.22
isort==4.3.21
jdcal==1.4.1
jieba==0.39
Jinja2==2.10.3
jsonpath==0.82
jsonpickle==1.2
kiwisolver==1.2.0
kombu==4.6.10
lazy-object-proxy==1.4.1
markdown2==2.3.8
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.3.1
mccabe==0.6.1
middleware==1.2.3
mistune==0.8.4
monotonic==1.5
mysqlclient==1.4.2.post1
nginx==0.0.1
numpy==1.18.3
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.5
packaging==19.1
pandas==1.0.3
path==15.0.0
Pillow==5.4.1
pipreqs==0.4.10
psutil==5.7.0
pycparser==2.20
pycryptodomex==3.9.4
Pygments==2.4.2
pylint==2.3.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.2
pyte==0.8.0
python-alipay-sdk==2.0.1
python-dateutil==2.8.1
python-logstash==0.4.6
python-memcached==1.59
python-slugify==3.0.3
pytz==2019.2
PyYAML==5.2
raven==6.10.0
rcssmin==1.0.6
redis==3.4.1
requests==2.22.0
requests-toolbelt==0.9.1
rjsmin==1.1.0
send-email==201212
serializers==0.2.4
six==1.12.0
SQLAlchemy==1.3.17
sqlparse==0.3.1
tablib==2.0.0
text-unidecode==1.2
thefuck==3.30
typed-ast==1.3.1
Unidecode==1.1.1
urllib3==1.25.7
vine==1.3.0
wcwidth==0.1.9
webencodings==0.5.1
WeRoBot==1.9.0
Whoosh==2.7.4
win-unicode-console==0.5
wrapt==1.11.2
wxPython==4.0.0b2
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0
yarg==0.1.9
zipp==3.1.0
uWSGI==2.0.19.1